+++ src/compiler/tokens/statements.ts
@@ -11,7 +11,7 @@
  * @copyright 2021-2022 Luna Klatzer
  * @since 0.1.0
  */
-import { CompilableParseToken, eligibleParentToken } from "./parse-token";
+import { CompilableParseToken, eligibleChildToken, eligibleParentToken } from "./parse-token";
 import {
 	CompoundStatementContext,
 	ExpressionStatementContext,
@@ -22,6 +22,7 @@ import {
 import { KipperProgramContext } from "../program-ctx";
 import { ScopeVariableDeclaration } from "../logic";
 import { VariableDeclaration } from "./definitions";
+import { Expression } from "./expressions";
 
 /**
  * Every antlr4 statement ctx type
@@ -95,6 +96,13 @@ export abstract class Statement extends CompilableParseToken {
 	public get scope(): KipperProgramContext | CompoundStatement {
 		return this._scope;
 	}
+
+	/**
+	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
+	 */
+	public abstract translateCtxAndChildren(): Array<Array<string>>;
 }
 
 /**
@@ -109,6 +117,8 @@ export class CompoundStatement extends Statement {
 	 */
 	protected override readonly _antlrContext: CompoundStatementContext;
 
+	protected readonly _children: Array<Statement>;
+
 	private _localScope: Array<ScopeVariableDeclaration>;
 
 	constructor(
@@ -119,6 +129,14 @@ export class CompoundStatement extends Statement {
 		super(antlrContext, parent, scope);
 		this._antlrContext = antlrContext;
 		this._localScope = [];
+		this._children = [];
+	}
+
+	/**
+	 * The children of this parse token.
+	 */
+	public get children(): Array<Statement> {
+		return this._children;
 	}
 
 	/**
@@ -154,10 +172,15 @@ export class CompoundStatement extends Statement {
 
 	/**
 	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
 	 */
-	public translateCtxAndChildren(): Array<string> {
-		// TODO!
-		return [];
+	public translateCtxAndChildren(): Array<Array<string>> {
+		let childCode: Array<Array<string>> = [];
+		for (let child of this.children) {
+			childCode = [...childCode, ...child.translateCtxAndChildren()];
+		}
+		return [["{"], ...childCode, ["}"]];
 	}
 }
 
@@ -173,6 +196,8 @@ export class SelectionStatement extends Statement {
 	 */
 	protected override readonly _antlrContext: SelectionStatementContext;
 
+	protected readonly _children: Array<Statement>;
+
 	constructor(
 		antlrContext: SelectionStatementContext,
 		parent: eligibleParentToken,
@@ -180,6 +205,14 @@ export class SelectionStatement extends Statement {
 	) {
 		super(antlrContext, parent, scope);
 		this._antlrContext = antlrContext;
+		this._children = [];
+	}
+
+	/**
+	 * The children of this parse token.
+	 */
+	public get children(): Array<Statement> {
+		return this._children;
 	}
 
 	/**
@@ -199,8 +232,10 @@ export class SelectionStatement extends Statement {
 
 	/**
 	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
 	 */
-	public translateCtxAndChildren(): Array<string> {
+	public translateCtxAndChildren(): Array<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -218,6 +253,8 @@ export class ExpressionStatement extends Statement {
 	 */
 	protected override readonly _antlrContext: ExpressionStatementContext;
 
+	protected readonly _children: Array<Expression>;
+
 	constructor(
 		antlrContext: ExpressionStatementContext,
 		parent: eligibleParentToken,
@@ -225,6 +262,14 @@ export class ExpressionStatement extends Statement {
 	) {
 		super(antlrContext, parent, scope);
 		this._antlrContext = antlrContext;
+		this._children = [];
+	}
+
+	/**
+	 * The children of this parse token.
+	 */
+	public get children(): Array<Expression> {
+		return this._children;
 	}
 
 	/**
@@ -244,10 +289,15 @@ export class ExpressionStatement extends Statement {
 
 	/**
 	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
 	 */
-	public translateCtxAndChildren(): Array<string> {
-		// TODO!
-		return [];
+	public translateCtxAndChildren(): Array<Array<string>> {
+		let childCode: Array<string> = [];
+		for (let child of this.children) {
+			childCode = [...childCode, ...child.translateCtxAndChildren()];
+		}
+		return [[...childCode, ";"]];
 	}
 }
 
@@ -263,6 +313,8 @@ export class IterationStatement extends Statement {
 	 */
 	protected override readonly _antlrContext: IterationStatementContext;
 
+	protected readonly _children: Array<Expression>;
+
 	constructor(
 		antlrContext: IterationStatementContext,
 		parent: eligibleParentToken,
@@ -270,6 +322,14 @@ export class IterationStatement extends Statement {
 	) {
 		super(antlrContext, parent, scope);
 		this._antlrContext = antlrContext;
+		this._children = [];
+	}
+
+	/**
+	 * The children of this parse token.
+	 */
+	public get children(): Array<Expression> {
+		return this._children;
 	}
 
 	/**
@@ -289,8 +349,10 @@ export class IterationStatement extends Statement {
 
 	/**
 	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
 	 */
-	public translateCtxAndChildren(): Array<string> {
+	public translateCtxAndChildren(): Array<Array<string>> {
 		// TODO!
 		return [];
 	}
@@ -308,6 +370,8 @@ export class JumpStatement extends Statement {
 	 */
 	protected override readonly _antlrContext: JumpStatementContext;
 
+	protected readonly _children: Array<Expression>;
+
 	constructor(
 		antlrContext: JumpStatementContext,
 		parent: eligibleParentToken,
@@ -315,6 +379,14 @@ export class JumpStatement extends Statement {
 	) {
 		super(antlrContext, parent, scope);
 		this._antlrContext = antlrContext;
+		this._children = [];
+	}
+
+	/**
+	 * The children of this parse token.
+	 */
+	public get children(): Array<Expression> {
+		return this._children;
 	}
 
 	/**
@@ -334,8 +406,10 @@ export class JumpStatement extends Statement {
 
 	/**
 	 * Generates the typescript code for this item, and all children (if they exist).
+	 *
+	 * Every item in the array represents a single line of code.
 	 */
-	public translateCtxAndChildren(): Array<string> {
+	public translateCtxAndChildren(): Array<Array<string>> {
 		// TODO!
 		return [];
 	}